% Generated by roxygen2 (4.0.1): do not edit by hand
\name{json2XML}
\alias{json2XML}
\title{Convert a JSON string to XML}
\usage{
json2XML(jsonString)
}
\arguments{
\item{jsonString}{A character string of JSON text}
}
\description{
\code{json2XML}
Convert a JSON string into a well-formed, element-normal XML string.
}
\section{Value}{
 An XML string to be parsed containing the structured data from the JSON string.
}
\examples{
\dontrun{
library(RJSONIO)
library(XML)
jsonString <- toJSON(list(a=1, r = list(b = 2:3, c= 'd')), collapse = '')
xData <- json2XML(jsonString)
> xmlParse(json2XML(jsonString))
<?xml version="1.0"?>
<root>
 <r>
  <b>2</b>
  <b>3</b>
  <c>d</c>
 </r>
 <a>1</a>
</root>
}
}

