% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xml2JSON}
\alias{xml2JSON}
\title{Convert a XML string to JSON}
\usage{
xml2JSON(xmlString)
}
\arguments{
\item{xmlString}{A character string of XML text}
}
\description{
\code{xml2JSON}
Convert a well-formed (but not necessarily valid) XML string into a JSON string. Some information may be lost in this transformation because JSON is a data format and XML is a document format. XML uses elements, attributes, and content text, while JSON uses unordered collections of name/value pairs and arrays of values. JSON does not does not like to distinguish between elements and attributes. Sequences of similar elements are represented as JSONArrays. Content text may be placed in a "content" member. Comments, prologs, DTDs, and <[ [ ]]> are ignored.
}
\section{Value}{
 A JSON string to be parsed containing the structured data from the XML string.
}
\examples{
\dontrun{
library(XML)
library(RJSONIO)
fileURL <-  "http://api.sportsdatallc.org/nfl-t1/2013/REG/1/statistics.xml?api_key=4dhyq3f3rfkp2cbm4yqbcuag"
xData <- readLines(fileURL)
jData <- xml2JSON(paste(xData, collapse='')) # convert to JSON
out <- fromJSON(jData)
do.call(rbind.data.frame, out$games$game[[1]]$team[[1]][[c('defense', 'player')]])
}
}

